synchronize types, and s/curproc/p/ in one forgotten place

+++ sys/arch/mac68k/include/cpu.h
@@ -1,4 +1,4 @@
-/*	$NetBSD: cpu.h,v 1.58 1999/02/25 23:13:41 is Exp $	*/
+/*	$NetBSD: cpu.h,v 1.59 1999/02/26 22:37:58 is Exp $	*/
 
 /*
  * Copyright (c) 1988 University of Utah.
@@ -327,7 +327,7 @@ void	proc_trampoline __P((void));
 void	loadustp __P((int));
 
 /* sys_machdep.c */
-void	cachectl1 __P((int, vaddr_t, u_long, struct proc *));
+void	cachectl1 __P((unsigned long, vaddr_t, size_t, struct proc *));
 
 /* vm_machdep.c */
 void	physaccess __P((caddr_t, caddr_t, register int, register int));
